requirements.txt
fastapi
uvicorn
transformers==4.30.2
torch==2.0.1
Pillow
requests
beautifulsoup4
